Add a solution to Exercise 2.2.iii.
There is no double struck lowercase omega in Unicode. (In fact, nearly
no double struck Greek letters exist.) I ended up adding the required
glyph as part of a stylistic set to cheat it into existence. The glyph
is not well-designed, the font manipulation is badly executed, the idea
of using a stylistic set is bad, and it’s akward to actually use. But
it’s the best I have at the moment, and at least fits the font.